pyproject.toml
[project]
name = "med-deepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rank-bm25>=0.2.2",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "filelock>=3.13.0",
]
